requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
discord.py>=2.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

